Moved GetDesiredIndentation to a new service
